_Yehu

+++ include/MovingObjects/MovingObjects.h
@@ -14,10 +14,10 @@ public:
 	void updateSprite();
 	void setGraphics(const sf::Vector2u&, int, bool = false, float switchTime = 0.2f);
 	void setSpeed(float);
-	void setDirection(const sf::Vector2i&);
+	void setDirection(const sf::Vector2f&);
 	void undoMove();
 
-	sf::Vector2i getDirection() const;
+	sf::Vector2f getDirection() const;
 
 protected:
 	Animation& getAnimation();
@@ -25,6 +25,6 @@ protected:
 
 private:
 	Animation m_animation;
-	sf::Vector2i m_direction;
+	sf::Vector2f m_direction;
 	float m_speed;
 };
