+++ src/MovingObjects/Animate/LinkAndStates/LinkStandingState.cpp
@@ -31,7 +31,7 @@ void LinkStandingState::enter(Link& link){
         link.move();
     }
 
-    sf::Vector2i currentDirection = link.getDirection();
+    sf::Vector2f currentDirection = link.getDirection();
 
     if(currentDirection.x == 1)
     {
